perf(ble): 240-byte notify payload, continuous ADPCM stream, 16 kHz recording
Enlarge the BLE notify path so recordings transfer in far fewer chunks and
at full 16 kHz quality, and remove ADPCM block-boundary clicks.

Notify payload 58 -> 240 bytes (kRecChunkPayload):
- WB5MMG side: P2P notify GATT characteristic 64 -> 248 (p2p_stm.c), at_buffer
  and its extern 160 -> 560 (ble_at_server.c/main.c). ATT_MTU is already 251,
  so 243-byte notifies fit in one packet.
- Shared AT driver: str_received / buffer_rx sizing to 560 and the rx
  size/cursor to uint16_t (were uint8_t, capping AT lines at 255 chars),
  NOTIF_VAL/INDIC_VAL val_tab 64 -> 248 (stm32wb_at.c/.h/_ble.h,
  stm32wb_at_client.c client_buff_tx 160 -> 560).
- U585 side: bleAtBuffer 160 -> 560. Both firmwares rebuilt and reflashed
  (WB via SW4/SW5 switch).

Continuous ADPCM stream (fixes clicks): the recorder used to emit one
self-contained 4-byte-header block per 256-sample feed, resetting the
predictor every block and producing audible boundary discontinuities
(measured: 256-sample-boundary jumps median 838 vs in-block 0). Now it
writes one header at Start() and appends raw nibbles continuously
(adpcm_encode_nibbles / adpcm_decode_nibbles added; block adpcm_encode/decode
kept for the UART audio path). Boundary jumps are now 0.

Recording back to 16 kHz (no decimation): the 4x-smaller chunk count buys
back the 2x data of full-rate audio - a 3 s clip is ~69 chunks at 16 kHz vs
219 at the old 8 kHz / 58-byte scheme. WAV rate REC_SAMPLE_RATE 8000 -> 16000.

Verified on hardware, thorough suite (5 short + 2 long recordings): 7/7 PASS,
100% delivered by auto-push (0 gap-fill), decoded sample counts match REC_INFO
exactly, byte-identical on repeated REC_GET of the same seq, block-boundary
jump median 0 (clicks gone), saved WAV 0.0% saturated. IDLE<->ACTIVE resume
5/5, CRC 0. Existing adpcm unit test still PASSES.

+++ Drivers/BSP/Components/stm32wb_at/stm32wb_at.h
@@ -56,7 +56,7 @@ extern "C" {
 /* Exported macros -----------------------------------------------------------*/
 
 /* Exported functions --------------------------------------------------------*/
-uint8_t stm32wb_at_Init(uint8_t *buff_rx, uint8_t buff_rx_size);
+uint8_t stm32wb_at_Init(uint8_t *buff_rx, uint16_t buff_rx_size);
 uint8_t stm32wb_at_Received(uint8_t byte);
 uint8_t stm32wb_at_Process_rx_frame(char * str);
 uint8_t str_to_byte_tab(char *str_in, uint8_t *tab, uint8_t* tab_len);
